add kl annealing
