+++ src/scripts/run_dsprites.sh
@@ -9,7 +9,7 @@ torchrun --nproc_per_node=2 --master_port=29502 src/train.py \
     --batch_size 256 \
     --num_epochs 30 \
     --learning_rate 5e-4 \
-    --beta 8 \
+    --beta 3 \
     --z_dim 2 \
     --coff 0.5 \
     --pre_kl \
